% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era_download.R
\name{era.download}
\alias{era.download}
\title{Download the string character for specific variable of the AWS dataset}
\usage{
era.download(era.name = era.name)
}
\arguments{
\item{era.name}{a string character with the route of the AWS bucket and the
name of the variable to download. The string character could be the return of
get.name.vars function}
}
\description{
Download the string character for specific variable of the AWS dataset
}
