% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_name_vars.R
\name{get.name.vars}
\alias{get.name.vars}
\title{Complete list of era5 dataset from AWS services}
\usage{
get.name.vars(year, month)
}
\arguments{
\item{year}{integer or character of 4 digits indicating the year}

\item{month}{integer of 1 or 2 digits indicating the month}
}
\value{
a character with the variable names of year and month selected
}
\description{
Complete list of era5 dataset from AWS services
}
